requirements/production.txt
requests
django
dj-database-url
psycopg2
djangorestframework
markdown
django-filter
django-ipware
geoip2
